Add curve families wiki
This summarizes the Zulip thread on curve families and also provides links to authoritative sources for most of the curves.

It sadly lacks visual examples. The graph could be improved greatly by annotating it with additional information.

Also a flaw of this PR is that the math doesn't render. That should probably be fixed before commit.