Add GPU sorting article (#33)
* Add GPU sorting article

* A bit more detail on DeviceRadixSort

Analysis of the DeviceRadixSort code including critiques of correctness/portability.

* Address review feedback

A number of stylistic changes, and a rework of critiques of DeviceRadixSort; those have been moved to the PR adding them to UnityGaussianSplatting. In addition, there's a link to the Zulip thread.

* DeviceRadixSort subgroup size portability

Add a sentence explaining the current state of DeviceRadixSort with respect to portability to different subgroup sizes.