Minor tweaks
Fix run-on sentence (thanks Daniel). Make link to stroke expansion paper point primarily to arXiv.

We decided in office hours today to make the scope strictly the month of April. There will be major news in May, not least of which the 0.1 release, but there will be separate publicity around that.