Change `too_many_arguments` to `allow`. (#79)
Whenever we trigger the `too_many_arguments` lint we tend to just add an
exception instead of changing the code. The lint is just causing
friction.

Our current argument numbers are:

| body of work | counts |
| -- | -- |
| xilem_web | 10 |
| kurbo | 8,8 |
| parley | 9,8,8 |
| vello | 10,8,8 |
| vello_shaders | 11,10,9,9,9,9,8 |
| vello example | 10,9,9,9,7,7 |
| vello_svg example | 9,8 |
| velato | 8 |
| velato examples | 9,9 |
| interpoli | 8 |
| glazier | 12 |

We discussed this and [ran a poll on
Zulip](https://xi.zulipchat.com/#narrow/channel/419691-linebender/topic/Standard.20Lint.20set/near/484735271).
The solution is to just `allow` this lint.