Prune the canonical lint set (#88)
We've had complaints that the lint set is too strict, which I think is
fair. As such, I suggest that we remove some of the most pedantic lints.
It also adds a new set which we "want to" follow, but which don't block
PRs.
It's expected that we would work to not regress these between (for
example) crate releases.

Of note:
1) `allow_attributes` is removed from the default set. This doesn't mean
we shouldn't still aim to use `expect`, but having an allow is fine.
2) `default_trait_access` is newly *added* to the new "periodic"
category.
3) `missing_panics_doc` is removed entirely.
4) `wildcard_imports` is removed entirely