Make `clippy::shadow_unrelated` a periodical lint (#90)
This is quite unfortunate, because this lint can be useful. It can
however be a bit of a drag in test code (although I'd argue that it also
improves that)

The reason I'm suggesting removing it from the default set entirely is:
1) It really breaks with Xilem, because of e.g. `button`
2) It's a little-bit over-eager sometimes (e.g.
https://github.com/linebender/vello/blob/782fb5a3ea8dc06bac614790131c2695802766ae/examples/scenes/src/simple_text.rs#L222)

However, it does still raise things which should be fixed (e.g.
https://github.com/linebender/vello/blob/782fb5a3ea8dc06bac614790131c2695802766ae/examples/scenes/src/mmark.rs#L90)
so we should still run it.