Add backports wiki entry. (#109)
We're pretty solidly in the business of doing backports now, so I think
it's time we properly standardize it. It'll make understanding all of
this easier across repos, and the wiki entry will also make it easier to
do in an urgent situation.

I based this on how we've been doing it thus far, with no major changes.

Some things worth pointing out:

### Branch naming

We've historically used `latest`, `v0.4.x`, and `0.3.x`. While the last
two both make it clear what series they are for, I went with `v0.4.x`
(and `v4.x.x`) for the standard because:
* We get to have a more specific pattern in branch protection (`v*.x` vs
`*.x`)
* It's closer to the tags, so `v0.4.4` and `v0.4.5` will be in the
`v0.4.x` branch

### Branch protection

Standardizing will allows us to set up automated rule checking. We'll
allow commits to the backport branch via PR only. Review required. CI
pass required.

There are some nuances to note though.

In order to support optionally choosing **Rebase and merge** for the
cherry picked commits (*while not wanting it for everything else, like
potential custom new code*) we won't be using merge queues for the
backport branches. This should be fine, because the backport branches
are likely to see very low PR traffic. To ensure that the verification
is still the same, the backport branch PRs will be required to be
up-to-date. No stale base. Which should also be fine, because, again,
low PR traffic.

The only technically required CI job will be *formatting*, every other
job's result status needs to be verified by the humans involved. The
allows us to cut down on administrative work. By only requiring
*formatting* we can export/import a single universal backport branch
ruleset across repos. Otherwise we would need either per-repo rulesets
(like we already do for `main`) and constant backporting of all CI
changes to all backport branches, or alternatively, per-branch rulesets
to avoid having to backport all CI changes. None of it is very
appealing. So let's try human attention for now.