Remove `unexpected_cfgs` from the lint set (#117)
This is a fully backwards-compatible change, because `unexpected_cfgs`
is already warn by default

Going through the history, this was added to the lint set because:
- Xilem set the lint explicitly to support the cfgs it needed (FALSE and
tarpaulin)
- https://github.com/linebender/peniko/pull/47 copied the lints from
Xilem, even though it didn't need to configure unexpected_cfgs
- The original lint set PR was based on the set in Peniko

I've treated this as v6.1 due to the backwards compatibility. We can
make the next change be v7.0 or just v7 again.