A few updates to the lint set (#122)
Having used
[`clippy::default_trait_access`](https://rust-lang.github.io/rust-clippy/stable/index.html#default_trait_access)
as an editor lint, I'm really happy with it. It also only ever makes
extremely local suggestions (making addressing it when writing a change
trivial), and significantly improves readability in a lot of cases. It's
also something which I've now found myself suggesting regularly in PR
reviews, as better style.
As such, I'm proposing that we move this into the proper lint set.

Separately,
[`unused_qualifications`](https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-qualifications)
is good for consistency (if you see `Rect` and `kurbo::Rect` next to
each other, you might reasonably assume that they are different types).
As such, I'm also proposing that we move this into the "proper" lint set
out of the periodic lints.
This proposal is more marginal, and I'm happy to demote it back to a
periodic lint if there's pushback.

Thirdly, I'm proposing removed the editor lints as a class; I don't
think anyone ever used them, and the theory that you can make drive-by
style fixes in otherwise unrelated PRs was never sound.